# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AfriLance freelance marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles/me:
    get:
      operationId: getMyProfile
      tags: [profiles]
      summary: Get current user's profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Profile not found
    put:
      operationId: upsertMyProfile
      tags: [profiles]
      summary: Create or update current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertProfileBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profiles/{id}:
    get:
      operationId: getProfile
      tags: [profiles]
      summary: Get a profile by user ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Not found
  /profiles/freelancers:
    get:
      operationId: listFreelancers
      tags: [profiles]
      summary: List freelancer profiles with filters
      parameters:
        - name: skill
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: minRate
          in: query
          schema:
            type: number
        - name: maxRate
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
        - name: featured
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of freelancers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FreelancerListResponse"

  # Jobs
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List job listings with filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: skill
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: remote
          in: query
          schema:
            type: boolean
        - name: minBudget
          in: query
          schema:
            type: number
        - name: maxBudget
          in: query
          schema:
            type: number
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
        - name: budgetType
          in: query
          schema:
            type: string
            enum: [fixed, hourly]
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [newest, oldest, budget_asc, budget_desc]
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Post a new job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateJobBody"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateJobBody"
      responses:
        "200":
          description: Job updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResponse"
components:
  schemas:
...
    Profile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        role:
          type: string
        name:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        hourlyRate:
          type: ["number", "null"]
        category:
          type: ["string", "null"]
        portfolioItems:
          type: array
          items:
            $ref: "#/components/schemas/PortfolioItem"
        avgRating:
          type: ["number", "null"]
        totalReviews:
          type: integer
        featured:
          type: boolean
        isBlocked:
          type: boolean
        isVerified:
          type: boolean
        isTopRated:
          type: boolean
        resumeUrl:
          type: ["string", "null"]
        verificationDocUrl:
          type: ["string", "null"]
        verificationStatus:
          type: string
        completedJobs:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - userId
        - role
        - name
        - skills
        - portfolioItems
        - totalReviews
        - featured
        - isBlocked
        - isVerified
        - isTopRated
        - verificationStatus
        - completedJobs
        - createdAt
    PortfolioItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        projectUrl:
          type: ["string", "null"]
      required:
        - id
        - title
    UpsertProfileBody:
      type: object
      properties:
        role:
          type: string
        name:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        hourlyRate:
          type: ["number", "null"]
        category:
          type: ["string", "null"]
        portfolioItems:
          type: array
          items:
            $ref: "#/components/schemas/PortfolioItem"
        resumeUrl:
          type: ["string", "null"]
        verificationDocUrl:
          type: ["string", "null"]
      required:
        - role
        - name
    FreelancerListResponse:
      type: object
      properties:
        freelancers:
          type: array
          items:
            $ref: "#/components/schemas/Profile"
        total:
          type: integer
      required:
        - freelancers
        - total
    Job:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        skills:
          type: array
          items:
            type: string
        budgetMin:
          type: ["number", "null"]
        budgetMax:
          type: ["number", "null"]
        budgetType:
          type: string
        location:
          type: ["string", "null"]
        remote:
          type: boolean
        deadline:
          type: ["string", "null"]
        status:
          type: string
        isFlagged:
          type: boolean
        proposalCount:
          type: integer
        clientName:
          type: ["string", "null"]
        clientAvatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - clientId
        - title
        - description
        - category
        - skills
        - budgetType
        - remote
        - status
        - isFlagged
        - proposalCount
        - createdAt
        - updatedAt
    CreateJobBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        skills:
          type: array
          items:
            type: string
        budgetMin:
          type: ["number", "null"]
        budgetMax:
          type: ["number", "null"]
        budgetType:
          type: string
        location:
          type: ["string", "null"]
        remote:
          type: boolean
        deadline:
          type: ["string", "null"]
      required:
        - title
        - description
        - category
        - skills
        - budgetType
        - remote
    UpdateJobBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        skills:
          type: array
          items:
            type: string
        budgetMin:
          type: ["number", "null"]
        budgetMax:
          type: ["number", "null"]
        budgetType:
          type: string
        location:
          type: ["string", "null"]
        remote:
          type: boolean
        deadline:
          type: ["string", "null"]
        status:
          type: string
    JobListResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
      required:
        - jobs
        - total
...
    RequestUploadUrlBody:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
      required:
        - name
        - size
        - contentType
    RequestUploadUrlResponse:
      type: object
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          type: object
          properties:
            name:
              type: string
            size:
              type: integer
            contentType:
              type: string
          required:
            - name
            - size
            - contentType
      required:
        - uploadURL
        - objectPath
        - metadata
...